apps/test_core_python.py
import unittest
import pandas as pd
from core_python import fetch_data, process_data, analyze_data, visualize_data, store_data

class TestWorkflowFunctions(unittest.TestCase):

    def setUp(self):
        # Setup a sample dataframe for testing
        self.sample_data = {
            'cogs': [100, 200, 300, 400, 500],
            'gross margin percentage': [5, 10, 15, 20, 25],
            'gross income': [5, 20, 45, 80, 125]
        }
        self.df = pd.DataFrame(self.sample_data)

    def test_fetch_data(self):
        df = fetch_data()  # Assuming fetch_data reads data from a file
        self.assertFalse(df.empty, "Fetch data failed")

    def test_process_data(self):
        processed_data = process_data(self.df)
        self.assertFalse(processed_data.empty, "Process data failed")
        self.assertEqual(len(processed_data), len(self.df.drop_duplicates()), "Process data did not remove duplicates correctly")

    def test_analyze_data(self):
        fetch_data_result = fetch_data()  # Assuming fetch_data is the first step
        processed_data = process_data(fetch_data_result)  # Process the fetched data
        analysis_result = analyze_data(processed_data)  # Analyze the processed data
        self.assertTrue(isinstance(analysis_result, pd.DataFrame), "Analyze data did not return a DataFrame")
        self.assertIn('cogs', analysis_result.columns, "Analysis result missing 'cogs' column")

    def test_visualize_data(self):
        # No direct assertion possible for visualization, assuming visualization works without errors
        # Consider how to handle visualization output if it's required for testing
        pass

    def test_store_data(self):
        store_result = store_data(self.df)  # Assuming store_data function writes data to a file
        self.assertEqual(store_result, "Data stored successfully", "Store data failed")

if __name__ == '__main__':
    unittest.main()
